Update lynis cron job to first do security updates
We are seeing a large number of lynis failures because of insecure
packages. To address, we run `unattended-upgrade` to install security
patches everytime before we run the lynis audit.

Additionally, we change the cron time to `*:30` so it does not conflict
with the nightly `unattended-upgrades` run we do at midnight.